(CU-85zt0qwa2 ots) refactor: global -> walk_handler, add helpers
