+++ addon/operators/walk_handler.py
@@ -29,8 +29,6 @@
 
 from time import time
 
-from os import path as os_path
-
 from bpy import ops as bpy_ops
 
 from bpy.types import Operator
@@ -41,94 +39,8 @@ from bpy.props import (
     BoolProperty,
 )
 
-from bpy_extras.io_utils import ImportHelper
-
 _walk_handler_invoked = False
 _walk_handler_invoke_time = 0
-
-
-class BM_OT_Global_Free_Icons(Operator):
-    """
-    Remove Preview Collections of custom icons initialized in the Global
-    Properties.
-    """
-
-    bl_idname = 'bakemaster.global_free_icons'
-    bl_label = "Remove Custom Icons"
-    bl_options = {'INTERNAL'}
-
-    def execute(self, context):
-        import bpy.utils.previews
-        preview_collections = context.scene.bakemaster.get_icon(reg=True)
-
-        for pcoll in preview_collections.values():
-            bpy.utils.previews.remove(pcoll)
-
-        preview_collections.clear()
-        return {'FINISHED'}
-
-
-class BM_OT_Global_Help(Operator):
-    bl_idname = 'bakemaster.global_help'
-    bl_label = "Help"
-    bl_description = "Press to visit the according BakeMaster's online documentation page"  # noqa: E501
-    bl_options = {'INTERNAL'}
-
-    page_id: StringProperty()
-    addon_version = "latest"
-
-    def get_url(self):
-        base = "https://bakemaster-blender-addon.readthedocs.io/en/"
-        urls = {
-            '': r'%s/',
-            'bakejobs': r'%s/',
-            'objects': r'%s/',
-            'maps': r'%s/',
-            'bake': r'%s/',
-            'bakehistory': r'%s/',
-        }
-
-        url = urls.get(self.page_id, r'%s/')
-        return url % (base + self.addon_version)
-
-    def invoke(self, context, _):
-        self.addon_version = context.scene.bakemaster.get_addon_version()
-        return self.execute(context)
-
-    def execute(self, _):
-        from webbrowser import open as webbrowser_open
-        webbrowser_open(self.get_url())
-        return {'FINISHED'}
-
-
-class BM_OT_Helper_FileChooseDialog(Operator, ImportHelper):
-    bl_idname = 'bakemaster.helper_filechoosedialog'
-    bl_label = "Choose a filepath"
-    bl_description = "Open a file browser, hold Shift to open the file, Alt to browse containing directory"  # noqa: E501
-    bl_options = {'INTERNAL'}
-
-    prop_name: StringProperty(options={'HIDDEN'})
-
-    config_lookup: BoolProperty(default=False, options={'HIDDEN'})
-    config_action: StringProperty(options={'HIDDEN'})
-
-    message: StringProperty(options={'HIDDEN'})
-
-    def process_exit(self):
-        if self.message != "":
-            self.report({'INFO'}, self.message)
-        return {'FINISHED'}
-
-    def execute(self, context):
-        if os_path.isfile(self.filepath):
-            self.filepath = os_path.dirname(self.filepath)
-        # no safe check if property is invalid
-        setattr(context.scene.bakemaster, self.prop_name, self.filepath)
-
-        if not self.config_lookup:
-            return self.process_exit()
-
-        bpy_ops.bakemaster.config('EXEC_DEFAULT', action=self.config_action)
 
 
 class BM_OT_Global_WalkHandler(Operator):
@@ -962,30 +874,4 @@ class BM_OT_Global_WalkData_Move(Operator):
     def invoke(self, context, _):
         if self.new_index > self.index:
             self.new_index -= 1
-        return self.execute(context)
-
-
-class BM_OT_Global_UI_Prop_Relinquish(Operator):
-    bl_idname = 'bakemaster.global_ui_prop_relinquish'
-    bl_label = "Set similar values"
-    bl_description = "Each selected item will have the same value of this property"  # noqa: E501
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    data_name: StringProperty()
-    prop_name: StringProperty()
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        data, _, attr = getattr(
-            bakemaster, "get_active_%s" % self.data_name)(bakemaster)
-        if data is None:
-            bakemaster.log("o0x0001", self)
-            return {'CANCELLED'}
-
-        container = getattr(bakemaster, "get_%s" % attr[:-1])(data)
-        setattr(container, self.prop_name, getattr(container, self.prop_name))
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
         return self.execute(context)
